+++ lib/config/theme/widgets/buttons/buttons.dart
@@ -1,2 +1,3 @@
 export 'app_header_action_button.dart';
 export 'app_primary_button.dart';
+export 'app_secondary_button.dart';