+++ lib/features/settings/application/historical_archives_workflow_panel_model_provider.g.dart
@@ -27,7 +27,7 @@ final historicalArchivesWorkflowPanelModelProvider =
 typedef HistoricalArchivesWorkflowPanelModelRef =
     AutoDisposeProviderRef<HistoricalArchivesWorkflowPanelViewModel>;
 String _$historicalArchivesWorkflowHash() =>
-    r'15339f2a79ea3edcdbd836fae66ba8b7e4f7ddfc';
+    r'3581f5f04b009bd79347b24b2808ca21ef577a22';
 
 /// See also [HistoricalArchivesWorkflow].
 @ProviderFor(HistoricalArchivesWorkflow)